[FIX] server.py: `resource` is not available on Windows
Rev. b29a5861e026742a5de0ee0f804af1008b17f322
introduces limits for the threadedserver

It uses `resource` to set the hard limit,
but unfortunately `resource` is not available on Windows.

As in the GeventServer, we therefore set the resource
hard limit within a `if os.name == 'posix'` block,
as in rev.
528a7cedfabed2d55c048007a9101c9d37197bab