[FIX] snailmail: instable behaviour on print
Since commit 82fcfdb231c6885a410821f143846136c3c813f9 we require a
certain key in the context to apply the custom snailmail report layout.
This was badly managed and would eventually confilt inside the modal
that links snailmail and the mail composer.

This fix stabilises the biehaviour by properly propagating the context
key.