[FIX] mrp_bom_cost: bom cost / bom quantity
Use case to reproduce:
- create a BoM of 50 finished products:
- Add a routing with an operation time of 1h
- Set a 1$/h cost on workcenter
- Set unit price from BoM

It will be 1$ for one product. It should be 0.02$
since it will take 1h for 50 products, so the price
should be 1/50

It happens because the system add the operation time by unit instead
of adding it only once.

related to commit f7ad5c26b62ba49904eaa3f43a6dffd368c3fc35

opw-1890005