[FIX] snailmail: avoid spam on user mailbox
the fix merged in 6fe33cdb3b0cd321dacaae7466c95c7de1b89be0
was only tackling half the problem. When you assign a user to an
activity, the said user will be notified by email. This also occur if
you change the assignee to the next activity even if you change it to
the same user.

This commit ensures that we only write on the user_id field when we
create the next activity and not during the update.