[FIX] sale_timesheet: UOM
Revert commits 57817d4a05ee0, f7369db6b861c5 and f1371e70793f598. They
introduce an incorrect UOM.

The original issue was an `AccessError` due to the fact that the
employee chosen was not in the same company than the user. Therefore,
instead of changing the logic of the UOM selection, we make sure to
select the employee in the appropriate company first, then fall back on
any company then.

opw-1888299