[FIX] website_event: Impossible to track proposal on website
Programming error introduced by cb1a6c007afd5829172c70d2023df5c7e6fac6c4

Steps to reproduce the bug:

- Create an event with category='conference'
- Checked Track Proposals on Website
- Save it

An error was raised saying that the website with the id of the event didn't exist.

Further, with the fix but without a website_id specified, checked Track proposal didn't
add the menu to track proposals on the conference.

opw:1904078

closes odoo/odoo#28688