[FIX] mail: show message document link in chat window
Revision on https://github.com/odoo/odoo/commit/90059730e9d9d078b883b7cb6828e2696e898e7d

The commit above consisted of refactoring the JS mail module.

A small regression was introduced on messages linked to a document
from a chat window: it was no longer displaying the link to redirect
to the document that this message originates from. This issue was
only isolated to chat windows: it was working fine in the Discuss app.

This commit re-introduces the displaying of the document link on
messages from chat windows.

Task-ID 1910119