[FIX] account: invoice sent when sent from list
The "sent" field on an invoice indicates if it has been sent to
customer or not.

When the invoice is sent to a customer from a form view the sent was
set, but when it was sent from the "Send" action on the list view,
it was not updated which seems wrong.

opw-1910937
closes #28796