[IMP] account: Add constraint on statement line currency
When importing a bank statement, we need to prevent the case that an inconsistent
currency_id is set on statement line (= statement currency).

E.g. Your company is EUR, your journal is EUR, your statement is imported with a csv having
currency_id = EUR.

The amount currency will be used in the reconciliation widget instead of the
amount. Furthermore, if you are not in the multi-currency group, the currency_id is invisible and
then, it becomes difficult to understand the reconciliation propositions in the widget.

-task: 1909317