[FIX] website_event: website menu on event not working
Commit 60f454572 tried to fix the website event menu issue.
But the fix was wrong and fixed only the event having a website_id set.
After the 60f454572 fix, event with a website_id set are working as expected
but the ones without a website_id are not working.
Indeed, it will create a website.menu without a website_id which will result in
the menu being duplicated for each website (multi-website rule).
The required on website_id when activating website menu on event was trying to
hide that new issue but it was problematic since:
1. We were removing a feature for no reason as it would not be possible anymore
   to create a generic event (no website) with website menu activated.
2. The required would only work in form view, demo data or custo data loaded
   through xml without a website_id but having website menu activated would
   still bug and would not have a menu. Worst, the menu would be duplicated for
   every website and added to the website navbar.

Now, regardless of the event having a website_id set or not, activating or
deactivating event menu will work.

closes odoo/odoo#29036