[FIX] mrp: BOM recursion with different variants
Fine tuning of 92223b634405ef300e837dd0656481b930a73509.

Let's consider two variants V1 and V2

The function check_cycle was too restrictive, it didn't allow
to create a BOM for V1 with a dependence to V2.

But a common use case could be for example:

A black T-shirt made  from a white T-shirt and black color.

opw:1911622

closes odoo/odoo#29075