[FIX] stock: wrong location technical field
The commit 7a76b1f2156ba7869d40834b0497935caa10990c changes the odoo object
of "rules" from stock.location.path to stock.rule. The field on the first
object to store the destination location was location_dest_id but is location_id
in the second one.

This commit fix the variable name that refers to the destination location

opw : 1919655

closes odoo/odoo#29755