[FIX] account: apply currency on uom change
When creating a new line on a customer invoice, we could get the price
unit in the company currency instead of the invoice currency.

This is because the uom_id onchange would ignore currency when setting
the price.

Changing an invoice line's price unit on uom_id change was introduced
in february 2018 in saas-11.3 with c3318056f.

Change to how currency is applied when setting an invoice line's price
unit was introduced in june 2018 in 9.0 with bf47ba3f.

So there was a missing part in saas-11.3 to adapt bf47ba3f for c3318056f
change that is added with this commit.

Without the fix, the added test failed with:

  AssertionError: 10.0 != 20

opw-1920504
closes #29776