[FIX] web: add tests for calendar start of the week and week numbering
The week_start has been added as an option since commit 5995c23cbf07be1.
This stores as a database parameter the day starting the week, [1.. 7] starting
on Monday, while the calendar options expect to get Sunday as 7 % 7 = 0.
As a result the week number would be shifted.

Commit d1be21f910db6702654c7453835a640d6ad2d320 solved the issue,
this commit adds tests for it.

opw 1916666

closes odoo/odoo#29838