[IMP] mail: restore support for all report_types in mail attachments
Commit 4b5eb1a9 introduced a new ir.action.report type: qweb-text.
Commit 899efae introduced an explicit check on the type, for html or pdf.
This removed the ability to support arbitrary reports through the call to the
generic render method, which dispatches to the correct one for a report_type.
We add support for the text format following the same logic.

The explicit error will be raised if the rendering fails,
e.g. in the case where the render method is not defined.

Thanks to @gdgellatly for reporting.
opw 1966208
closes #28513
closes #30013
closes #32532

closes odoo/odoo#32773

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>