[ADD] test_inherits_depends: new test module
This commit adds a test to test the scenario fixed in the parent commit.

Creating a new test module is necessary as the bug is only revealed when the
current module is different than the model._original_module (cf 484e4d59d19)

Before that commit, was failing only the second search of
test_ir_model_data_inherits_depends.
Add other test to avoid regression of the bugs corrected at 3afd450c and 484e4d5

closes odoo/odoo#30116