[FIX] account: Date maturity in duplicate journal entry.
Before this commit, when the journal entry date was duplicated, the
maturity date (in the journal entry lines) stayed unchanged (it kept the
values of the original journal) and the user wasn't able to modify it.

Now, when the user duplicated a journal entry, the user can modify the
maturity date.

Note that, as is the case for the journal entry date, the maturity date
can't be modified after saving the journal entry.

OPW-1914706

closes odoo/odoo#30393