[FIX] base: partner search no alnum->no VAT match
When we search a partner, there is a system that show partner who match
exactly the searched term with non-alphanumeric character removed.

But since 8915b66cec there is an issue that wrong partner match if:

- they have an empty string as VAT in the database (odd but can happen)
- if the searched term only contains non-alphanumberic characters

eg. i search 'é' -> all partner with empty string vat are found since
    the empty search '' matches empty VAT ''

With this changeset, if there is no alphanumeric character we will check
depending on the operator `vat=NULL`, `vat like NULL`, `vat ilike NULL`
that all will not match any result (and in the query plan does not take
time).

opw-1929563
closes #30732