[FIX] website_sale_stock: no edit out of stock msg
On a product page on the website, with website_sale_stock there may be
an information about the product being out of stock or not.

This information is loaded on the fly for the current variant (switching
variant will reload it) and so this should not be modified with the
editor because:

- the saved edition would appear on all product (even if those in stock)
- the saved edition is replace quickly with the real state

So if we want to change the content, there is no other way than a custom
change on JS template website_sale_stock.product_availability or
modifying translations.

opw-1936179
closes #31135