[FIX] sale_stock: display picking in portal
When a picking is linked to a sale order, it is not displayed anymore
ont the user portal.

This comes from a fix in rev: https://github.com/odoo/odoo/commit/9e203f6371134ced086ea9382680ad5c3568269d
that hides the shipping_adress div, which is the one targeted by the
xpath for shoing picking. To fix this behavior, we changed the xpath.

TASK-ID: 1939813

closes odoo/odoo#31267