[FIX] product, stock_account: cost currency
The cost of the product is a property field, meaning that:
- its value depends on the current company;
- its currency is the company currency

However, several views use the monetary widget to display its currency
with the default `currency_id` field. This can be incorrect, since the
latter is computed based on the product's company.

We introduce the new computed field `cost_currency_id` which has the
appropriate value and can be used in the views.

Closes #30616
opw-1934443