[REV] models: do not erase master version
This commit has been backported in 11.0 [1] and then reverted [2],
because it introduced a regression in web_studio, discovered at opw-1947380

As no forward-port has been done between the backport and the revert,
the forward-port of the revert also contains the patch itself, resulting
in a noop.

Actually revert c6ad989ed060e87a48e7907c3358e2462f4212e4 in branch `12.0`.

[1] c6bfe99698169bb0d8d1dc8f3f0040e8cc2fcd2c
[2] 7e36b4d547534e96a5a58d1b334f29e3bafb340c

closes odoo/odoo#31637

Signed-off-by: Christophe Simonis <chs@odoo.com>