[FIX] account: invoice alias in multi company
Install accounting in a multi company environment.
Configure email server.
Configure an alias for the vendor bill journal per company.

Before this commit, all the invoices were created in the same company,
regardless of the alias used.

Now, the invoice is created in the company of the used alias.

opw-1939909

closes odoo/odoo#31948

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>