[FIX] point_of_sale: crashing tax cash basis when no reconciled order line
Because of rev. c348d3381178f8bb0c1a4a77488a8c62a6ca1f85
when none of the pos order lines were reconciled,
the closing of the point of sale sessions failed
because of the tax cash basis entry generation.

There is no need to create the tax cash basis entries as nothing has been reconciled.

A known case is when the the bank journal credit account is set to a receivable account,
which has as effect to fully reconcile the payment line with its counterpart,
leaving no payment lines to reconcile with the order debit lines.
This is obviously not a correct accounting configuration,
but the point of sale validation did no fail because of this previously.

Fixes #32330

closes odoo/odoo#32399

Signed-off-by: Denis Ledoux <beledouxdenis@users.noreply.github.com>