[FIX] l10n_fr_pos_cert: fpos settings
The computation of hash integrity depends on the field
`tax_ids_after_fiscal_position`. However, this field is computed, but
not stored.

It means that if one modifies a fiscal position used in the POS, the
computation of `tax_ids_after_fiscal_position` will change.
Consequently, the hash computation of the order will be modified, and
the hash integrity of the journal will be corrupted.

We prevent the modification of the taxes implied in a fiscal position if
any POS order use it.

Closes #32665
opw-1969009

closes odoo/odoo#32701

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>