[FIX] website_sale: unit price
Commit e5a6850ab2032cfd54a018da introduced an error in the following
case:
- Go to Sales > Configuration > Settings
- Enable "Multiple Sales Prices per Product" and select option "Prices
  computed from formulas (discounts, margins, roundings)"
- Go to Sales > Catalog / Pricelists
- Select "Public Pricelist" and set "Discount Policy: Show public price
  & discount to the customer"
- Define the pricelist as based on the cost of the product with a
  negative discount for a product P
- Go to the eCommerce, buy the product

The price used in the cart is the cost price instead of the cost +
negative discount.

This is because we force the discount to be zero in this case while
using the price of the product without the pricelist.

We usually don't want to show a negative discount, since it means the
client pays more than usual. However we still want to use the price
defined on the pricelist. We modify as follow: if the discount is
negative, we don't show it to the user (value = 0), but we use the unit
price with the discount included.

opw-1967870

closes odoo/odoo#32768

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>