[FIX] website_sale_wishlist, js: add to cart qty
The selector used for finding the qty field is simply 'qty', which is useless
in most cases.

While there is actually no quantity input in a wishlist template, it could be
leveraged by extension modules.

Support for same markup as found in product page seems appropriate:
input[name="add_qty"]

closes odoo/odoo#33224

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>