[FIX] stock_warehouse: _get_all_routes() override
Usecase to reproduce:
- Install stock, mrp, purchase
- Go on warehouse form view
- Click on Routes button
Either one of the route manufacture or buy is missing.

All implementations/overrides of stock.warehouse._get_all_routes()
call their parent's implementation of get_all_routes_for_wh().
which is a reference to stock.warehouse._get_all_routes() but that means
that the route 'mixin' behavior can be skipped.
(the reference is to the base implementation of _get_all_routes())

Fixes #33286
Fixes #33288

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>