[FIX] mail: remove readonly on recommended next activity
Fine-tuning of cb8e2aaf6cb8f4ac34810a9dd5b0ad09276bb2cf.
It does not make sense to be able to write on mail.activity.type
using the related fields defined on mail.activity.

But this field is not a related one, and it is necessary to be writable to make
the 'recommended next activity' feature work at all.

opw 1974236

closes odoo/odoo#33325

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>