[FIX] crm: set the email_from from the kanban quick create
When setting the partner of a lead, on the form view the email_from
is automatically populated by the onchange on the partner_id.
This is not the case when created programmatically,
and this is not even the case when created from the kanban view.

This can lead to an unexpected failure to send emails down the line.

We apply the onchange in the create to fill all values consistently in all cases.
In case the values were already set, this simply won't change anything.

Note that we need to adapt the test from website_crm_partner_assign.
It uses demo data that makes little sense.
The opportunity crm.crm_case_21 is set in the UK,
with a partner base.res_partner_4 in the US, in the state Escuintla of Guatemala.
With the change from above, all the address fields on the opportunity are
populated by the onchange, from the partner.

opw 2000241

closes odoo/odoo#33622

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>