[FIX] sale,product_email_template: match the help with the behaviour
Before this commit, the help of the product_email_template module said
that the email is sent when the invoice is paid. This is not the case,
the current behaviour of the module is that the email is sent when the
invoice is validated.

Fine-tuning of : f20e3621a55af67c980e4328e38a729617010d1d,
815d20e2586dda77bb6b2888a42579038afb7ca2
and 2f59f6a7ee387c6cea5a9c4ff7dcc9d8964a5575

opw-1973946

closes odoo/odoo#33686

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>