[FIX] website_sale_delivery: restore monetary widget
In Odoo 11.0, 7139f0cc8691 fixed some flows related to prices. By doing that,
the `t-options` dict was set on a variable to avoid code duplication.

But starting from Odoo 12.0, setting a variable on `t-options` attribute seems
to do nothing and is ignored.

This commit fix it by setting the price rather than the options on a variable.

Fixes #33451

closes odoo/odoo#33900

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>