[FIX] website_sale: fix modal price when Select Quantity is disabled
To reproduce:

- Go on the website page of a product with optional products.
- Disable the Select Quantity customize show.
- Click add to cart.

This will open the optional product modal, where the quantity will be displayed.

However changing that quantity will have no effect, especially the price will
not update accordingly.

The real issue is that the quantity should not be displayed there if Select
Quantity is disabled, but this has never worked before, so this is outside the
scope of this fix.

Anyway, since the quantity is present there, we would at least expect it to
correctly recompute the price, which worked fine when the modal was introduced,
but has been broken since 747be0175dc9cc9cf65c55b74048d95282885c02

opw-2005317

closes odoo/odoo#33992

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>