[FIX] mrp: wrong value on cost report
Follow up to:
- commit b2bb5ef5c4ee5a7244a8048a8b346a445c52d6dc
- commit 782620043197107d69c0cf18ce42ba54fe08e6b9
- commit f22f012ad6747e1ece92f7a0679cc70ba8c818dd

In order to be able to edit a done raw move after the production
and having a correct cost report with consistent values. The
simple rule that a done move should have its quantity equals to
its initial demand. Since report is based on initial demand it
should follow this rule and it's also a good practice because a
lot of computation are based on it (easier than quantity done
because it's already set on the correct quant uom).

opw-2033259

closes odoo/odoo#34695

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>