[FIX] website_payment: consider authorized transactions as done
Forward port of 35ec5e921a8b10bdbac2fc610eecbd16dcd06a54

Before this commit, when a transaction has as status 'authorized', the
transactions' confirm page gives an error : 'Oops! There was a problem
with your payment.'

Now, the page gives the same message as when the transaction has the
status 'done' : 'Your payment was successful! It may take some time to
be validated on our end.'

opw-1984325