[FIX] stock: open quants in product
Before this commit, there was a different when opening the quants (the
'On Hand' button)  from the product or from the product template. Only
from the product template, the functions _merge_quants and
_unlink_zero_quants were called. This means that when opening from the
product, the quants were not merged, and the empty quants were not
removed.

Now, from the product or from the product template when opening the
quants both functions are called.

Fine-tuning of bc775148cc9dfbb79f834cc4f3bf0f947ba0d9f7

opw-2035001

closes odoo/odoo#34987

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>